Move AWS components to PSR4 directory
