+++ src/Aws/Eks/Detector.php
@@ -17,7 +17,7 @@ declare(strict_types=1);
  * limitations under the License.
  */
 
-namespace Detectors\Aws;
+namespace OpenTelemetry\Aws\Eks;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\RequestException;
@@ -31,7 +31,7 @@ use OpenTelemetry\SDK\Trace\Attributes;
  * Elastic Kubernetes and return a {@link Resource} populated with data about the Kubernetes
  * plugins of AWS X-Ray. Returns an empty Resource if detection fails.
  */
-class EksDetector
+class Detector
 {
     // Credentials and path for locating API
     private const K8S_SVC_URL = 'kubernetes.default.svc';
@@ -46,7 +46,7 @@ class EksDetector
     private $processData;
     private $guzzle;
     
-    public function __construct(EksProcessDataProvider $processData, Client $guzzle)
+    public function __construct(DataProvider $processData, Client $guzzle)
     {
         $this->processData = $processData;
         $this->guzzle = $guzzle;
@@ -162,39 +162,5 @@ class EksDetector
 
             return false;
         }
-    }
-}
-
-/**
- * Separated from above class to be able to test using
- * mock values through unit tests
- */
-class EksProcessDataProvider
-{
-    private const DEFAULT_CGROUP_PATH = '/proc/self/cgroup';
-    private const K8S_TOKEN_PATH = '/var/run/secrets/kubernetes.io/serviceaccount/token';
-    private const K8S_CERT_PATH = '/var/run/secrets/kubernetes.io/serviceaccount/ca.crt';
-
-    public function getK8sHeader()
-    {
-        $credHeader = file_get_contents(self::K8S_TOKEN_PATH);
-        
-        if ($credHeader) {
-            return 'Bearer' . $credHeader;
-        }
-
-        //TODO: Add log 'Unable to load K8s client token'
-        return null;
-    }
-
-    // Check if there exists a k8s certification file
-    public function isK8s()
-    {
-        return file_get_contents(self::K8S_CERT_PATH);
-    }
-
-    public function getCgroupData()
-    {
-        return file(self::DEFAULT_CGROUP_PATH, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
     }
 }
