+++ src/Aws/Xray/stacktraceTranslator.php
